refactor: Design Smell Refactoring (Tools.java) - Insufficient Modularization CodeSmell
